Fetch complainant details for document.

+++ src/main/java/uk/gov/digital/ho/hocs/cms/domain/repository/ComplaintsRepository.java
@@ -4,6 +4,8 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 import uk.gov.digital.ho.hocs.cms.domain.model.ComplaintExtractRecord;
 
+import java.math.BigDecimal;
+
 @Repository
-public interface ComplaintsRepository extends CrudRepository<ComplaintExtractRecord, Long> {
+public interface ComplaintsRepository extends CrudRepository<ComplaintExtractRecord, BigDecimal> {
 }
