+++ src/main/java/uk/gov/digital/ho/hocs/cms/domain/repository/DocumentsRepository.java
@@ -8,7 +8,7 @@ import uk.gov.digital.ho.hocs.cms.domain.model.DocumentExtractRecord;
 import java.math.BigDecimal;
 
 @Repository
-public interface DocumentsRepository extends CrudRepository<DocumentExtractRecord, Long> {
+public interface DocumentsRepository extends CrudRepository<DocumentExtractRecord, BigDecimal> {
 
     @Query(value = "SELECT count(*) FROM documents d WHERE d.case_id = ?1 AND d.document_extracted = 'false'", nativeQuery = true)
     BigDecimal findFailedDocumentsForCase(BigDecimal caseId);
